code refactor by each feature
