+++ lib/adopt/screens.dart/adopt_screen.dart
