+++ lib/common/models/breed_characteristic.dart
