+++ lib/common/widgets/text_highlight_widget.dart
