+++ lib/explore/widgets.dart/search_status_widget.dart
