+++ lib/explore/widgets.dart/search_widget.dart
