+++ lib/explore/widgets.dart/widgets.dart
@@ -1,0 +1,5 @@
+export './description_widget.dart';
+export './dog_tile_widget.dart';
+export './search_status_widget.dart';
+export './search_text_field.dart';
+export './search_widget.dart';
