+++ lib/main.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:puppy_world/providers/dog_tiles_provider.dart';
-import 'package:puppy_world/screens/main_tab_screen.dart';
-import 'package:puppy_world/theme/style.dart';
+
+import './theme/style.dart';
+import './common/screens/main_tab_screen.dart';
+import './explore/providers.dart/providers.dart';
 
 void main() async {
   runApp(PuppyWorld());
