--- lib/screens/screens.dart
@@ -1,5 +1,0 @@
-export 'explore/dog_details_screen.dart';
-export 'adopt/adopt_screen.dart';
-export 'explore/explore_screen.dart';
-export 'account/account_screen.dart';
-export 'explore/search_dogs_screen.dart';
